+++ core/dbt/flags.py
@@ -35,6 +35,7 @@ WHICH = None
 INDIRECT_SELECTION = None
 LOG_CACHE_EVENTS = None
 EVENT_BUFFER_SIZE = 100000
+QUIET = None
 
 # Global CLI defaults. These flags are set from three places:
 # CLI args, environment variables, and user_config (profiles.yml).
@@ -55,7 +56,8 @@ flag_defaults = {
     "PRINTER_WIDTH": 80,
     "INDIRECT_SELECTION": 'eager',
     "LOG_CACHE_EVENTS": False,
-    "EVENT_BUFFER_SIZE": 100000
+    "EVENT_BUFFER_SIZE": 100000,
+    "QUIET": False
 }
 
 
@@ -103,7 +105,7 @@ def set_from_args(args, user_config):
         USE_EXPERIMENTAL_PARSER, STATIC_PARSER, WRITE_JSON, PARTIAL_PARSE, \
         USE_COLORS, STORE_FAILURES, PROFILES_DIR, DEBUG, LOG_FORMAT, INDIRECT_SELECTION, \
         VERSION_CHECK, FAIL_FAST, SEND_ANONYMOUS_USAGE_STATS, PRINTER_WIDTH, \
-        WHICH, LOG_CACHE_EVENTS, EVENT_BUFFER_SIZE
+        WHICH, LOG_CACHE_EVENTS, EVENT_BUFFER_SIZE, QUIET
 
     STRICT_MODE = False  # backwards compatibility
     # cli args without user_config or env var option
@@ -128,6 +130,7 @@ def set_from_args(args, user_config):
     INDIRECT_SELECTION = get_flag_value('INDIRECT_SELECTION', args, user_config)
     LOG_CACHE_EVENTS = get_flag_value('LOG_CACHE_EVENTS', args, user_config)
     EVENT_BUFFER_SIZE = get_flag_value('EVENT_BUFFER_SIZE', args, user_config)
+    QUIET = get_flag_value('QUIET', args, user_config)
 
 
 def get_flag_value(flag, args, user_config):
@@ -179,5 +182,6 @@ def get_flag_dict():
         "printer_width": PRINTER_WIDTH,
         "indirect_selection": INDIRECT_SELECTION,
         "log_cache_events": LOG_CACHE_EVENTS,
-        "event_buffer_size": EVENT_BUFFER_SIZE
+        "event_buffer_size": EVENT_BUFFER_SIZE,
+        "quiet": QUIET
     }