+++ test/integration/044_run_operations_tests/macros/happy_macros.sql
@@ -49,3 +49,8 @@
     {{ log((node | string), info=True)}}
   {% endfor %}
 {% endmacro %}
+
+
+{% macro print_something() %}
+  {{ print("You're doing awesome!") }}
+{% endmacro %}